services/bess_map/requirements.txt
numpy
sqlalchemy
pulp
openpyxl
altair
streamlit==1.43.2
pandas==2.2.3
plotly==6.0.1
psycopg2-binary==2.9.10
openai==1.68.2
requests==2.32.3
reportlab==4.3.1
python-dotenv==1.0.1
boto3